created filter for today, and search with name

+++ Controllers/helsinkiApiController.js
@@ -10,11 +10,38 @@ const update = async (req, res) => {
     res.json(data[0]);
 }
 
-const getAll = async (limit) => {
+const getAll = async (limit, today, nameIncludes) => {
+
+    if(today && today === true) {
+        let data = await helsinkiModel.find({"event_dates.starting_day":{ 
+            $lte: new Date(new Date().getTime() + 7 * 60 * 60 * 24 * 1000).toISOString(),
+            $gte: new Date(new Date().getTime() - 1 * 60 * 60 * 24 * 1000).toISOString()
+            
+        }
+        }).limit(3);
+        return data;
+    }
+
+    if(nameIncludes) {
+        let data = await helsinkiModel.find( {$and: [{"event_dates.starting_day":{ 
+            $lte: new Date(new Date().getTime() + 16 * 60 * 60 * 24 * 1000).toISOString(),
+        }},{"name.fi":{
+            $regex: nameIncludes 
+        }}
+    ]}).limit(20);
+        return data;    
+    }
+
     //Find events that start 2 weeks into the future -> this because we will only also to show them weather included
     let data = await helsinkiModel.find({"event_dates.starting_day":{ 
         $lte: new Date(new Date().getTime() + 16 * 60 * 60 * 24 * 1000).toISOString()}
     }).limit(limit ? limit : 10);
+    return data;
+}
+
+const getOne = async (name) => {
+    //getOne from the database with the id
+    let data = await helsinkiModel.find({"id": name});
     return data;
 }
 
@@ -37,24 +64,32 @@ const createModel = (item) => {
     return eventModel;
 }
 
-const DeleteOldOnes = async () => {
+const DeleteOldOnes = async (req,res) => {
     let data = await helsinkiModel.find({});
-    data.map(async (event) => {
-        let ts = Date.now();
+    let deletedID = 0;
+
+    //TEST this one
+    await data.map(async (event) => {
+        let ts = new Date(new Date().getTime() * 1000).getTime()
+        console.log(event.event_dates.starting_day.getTime() +"<<<<"+ ts)
         if (!event.event_dates.starting_day) {
+            deletedID++;
             await helsinkiModel.deleteOne({
                 id: event.id
             })
-        } else if (event.event_dates.starting_day.getTime() < ts) {
+        } else if (event.event_dates.starting_day.getTime() < ts && event.event_dates.ending_day.getTime() < ts) {
+            deletedID++;            
             await helsinkiModel.deleteOne({
                 id: event.id
             })
         }
     })
+    res.json("Number of deleted items:"+ deletedID)
 }
 
 module.exports = {
     update,
     DeleteOldOnes,
-    getAll
+    getAll,
+    getOne
 }
